requirements.txt
# Core
streamlit>=1.55.0
yfinance>=1.2.0
plotly>=6.0.0
pandas>=2.3.0
numpy>=2.4.0
python-dotenv>=1.0.0

# Future (uncomment when needed)
# fredapi>=0.5.0          # Macro data (needs FRED API key)
# edgartools>=0.30.0      # SEC EDGAR filings
# anthropic>=0.50.0       # Claude API (needs API key)
# openai-whisper>=20230918 # Earnings call transcription
# scipy>=1.15.0           # Statistical calculations
# openpyxl>=3.1.0         # Excel export
# weasyprint>=62.0        # PDF export
